Move admin actions into issue box
