+++ client/src/components/Admin/Home.js
@@ -1,17 +1,15 @@
 import React from 'react';
-import { IssueAdminContainer } from './IssueAdmin';
+import { IssueContainer } from './Issue';
 import Alternatives from './Alternatives';
 import { ConcludedIssueListContainer } from '../ConcludedIssueList';
 import IssueStatus from '../IssueStatus';
-import Pin from './Pin';
 
 
 const AdminHome = () => (
   <div>
     <div className="ActiveIssue-components">
       <div className="ActiveIssue-Vote-wrapper">
-        <Pin code="DEADBEEF" />
-        <IssueAdminContainer />
+        <IssueContainer />
         <Alternatives />
       </div>
       <IssueStatus />
