+++ client/src/components/Admin/Pin.js
@@ -1,5 +1,4 @@
 import React, { PropTypes } from 'react';
-import '../../css/Pin.css';
 
 const Pin = ({ code }) => (
   <div className="Pin">
