--- client/src/css/Pin.css
@@ -1,5 +1,0 @@
-.Pin {
-    font-size: 1.1em;
-    color: #333;
-    margin: 2rem 0;
-}
